refractor: pdfgeneration functions -> pdfGenerationUtil.ts
